requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.1.0=pyhd8ed1ab_0
aiohttp=3.8.3=py310h5764c6d_0
aiosignal=1.2.0=pyhd8ed1ab_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py_1
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.9.14=py310h06a4308_0
cffi=1.15.0=py310hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py310hff52083_0
cryptography=37.0.1=py310h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py310h06a4308_0
etils=0.6.0=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
ffmpeg=4.2.2=h20bf706_0
flatbuffers=2.0.0=h2531618_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.1=py310h5764c6d_0
gast=0.5.3=pyhd3eb1b0_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.11.1=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.46.3=hbd84cd8_2
grpcio=1.46.0=py310h44b9e0c_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.7.0=py310he06866b_0
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.4=py310hff52083_0
importlib_resources=5.8.0=pyhd8ed1ab_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py310h06a4308_0
ipython=8.3.0=py310h06a4308_0
jax=0.3.14=pyhd8ed1ab_1
jaxlib=0.3.14=cpu_py310hacd7525_0
jedi=0.18.1=py310h06a4308_1
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jupyter_client=7.2.2=py310h06a4308_0
jupyter_core=4.10.0=py310h06a4308_0
keras=2.8.0=py310h06a4308_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.2=py310h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libcurl=7.84.0=h91b91d3_0
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=12.1.0=h8d9b700_16
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=12.1.0=h8d9b700_16
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h6239696_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=12.1.0=ha89aaad_16
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
libzlib=1.2.12=h166bdaf_1
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py310h5764c6d_1
matplotlib=3.5.2=py310hff52083_0
matplotlib-base=3.5.2=py310h5701ce4_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
multidict=6.0.2=py310h5764c6d_1
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
numexpr=2.8.3=py310hcea2de6_0
numpy=1.22.3=py310hfa59a62_0
numpy-base=1.22.3=py310h9585f30_0
oauthlib=3.2.1=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py310h6a678d5_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py310h22f2fdc_0
pip=21.2.4=py310h06a4308_0
pluggy=1.0.0=py310h06a4308_1
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py310hd8f1fbe_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py=1.11.0=pyhd3eb1b0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.5.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py310h295c915_6
pysocks=1.7.1=py310h06a4308_0
pytest=7.1.2=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=pyhd8ed1ab_0
python-flatbuffers=2.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.11.0=py3.10_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py310h06a4308_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py310h295c915_2
qt=5.9.7=h5867ecd_1
re2=2022.06.01=h27087fc_0
readline=8.1.2=h7f8727e_1
requests=2.28.0=py310h06a4308_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.0.1=py310h00e6091_0
scipy=1.7.3=py310hfa59a62_0
setuptools=61.2.0=py310h06a4308_0
sip=4.19.13=py310h295c915_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.38.5=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.8.0=py310h06a4308_0
tensorboard-data-server=0.6.0=py310h597c629_2
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.8.2=mkl_py310hd2b8f8c_0
tensorflow-base=2.8.2=mkl_py310hf890080_0
tensorflow-estimator=2.8.0=py310h2f386ee_0
termcolor=1.1.0=py310h06a4308_1
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=py310h06a4308_0
torchaudio=0.11.0=py310_cu102
torchvision=0.12.0=py310_cu102
tornado=6.1=py310h7f8727e_0
tqdm=4.64.1=py310h06a4308_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py310h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=py310h5eee18b_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7b6447c_0
yarl=1.7.2=py310h5764c6d_2
zeromq=4.3.4=h2531618_0
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h166bdaf_1
zstd=1.5.2=ha4553b6_0
